playground/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.3
python-dotenv==1.0.0
motor==3.3.2  # Async MongoDB driver

# Google Workspace Integration (Optional)
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Admin Dashboard (Optional)
streamlit>=1.29.0
pandas>=2.0.0
plotly>=5.18.0
